backend/requirements.txt
flask
flask-cors
pymupdf
pdf2image
pytesseract
Pillow
sentence-transformers
scikit-learn
numpy
